:bug:(heredoc): fixed variable expansion for quote-delimited limiter
Before, because we removed quotes for all token, the heredoc was
automatically expanding the variables, even if the limiter was quoted.
Now, we remove the quotes for the limiter later, inside the heredoc
function. This way, it doesn't expand the varibles when the limiter
is quoted, but we don't have to write the limiter with quotes to
close the heredoc file.